+++ tst/components/ui/Chat/ChannelList/index.test.tsx
@@ -6,27 +6,26 @@ import React from 'react';
 
 import ChannelList from '../../../../../src/components/ui/Chat/ChannelList';
 import ChannelItem from '../../../../../src/components/ui/Chat/ChannelList/ChannelItem';
-import lightTheme from '../../../../../src/theme/light'
+import lightTheme from '../../../../../src/theme/light';
 import { renderWithTheme } from '../../../../test-helpers';
 
 describe('ChannelList', () => {
-
   const defaultComponent = (
-    <ChannelList activeChannelId='abc' setActiveChannelId={() => null}>
-      <ChannelItem name='Channel 1' id='abc' />
+    <ChannelList>
+      <ChannelItem name="Channel 1" id="abc" onClick={() => null} />
     </ChannelList>
   );
 
   it('renders a ChannelList', () => {
     const component = defaultComponent;
-    const { getByTestId } = renderWithTheme(lightTheme, component)
+    const { getByTestId } = renderWithTheme(lightTheme, component);
     const el = getByTestId('channel-list');
     expect(el).toBeInTheDocument();
   });
 
   it('should render a ChannelItem child component', () => {
     const component = defaultComponent;
-    const { getByText } = renderWithTheme(lightTheme, component)
+    const { getByText } = renderWithTheme(lightTheme, component);
     const el = getByText('Channel 1');
     expect(el).toBeInTheDocument();
   });