+++ tst/snapshots/Chat/ChannelList/index.test.tsx
@@ -4,7 +4,7 @@
 describe('ChannelList', () => {
   it('visually looks correct', async () => {
     await page.goto(
-      'http://host.docker.internal:9009/iframe.html?id=ui-components-chat--channel-list&viewMode=story'
+      'http://host.docker.internal:9009/iframe.html?id=ui-components-chat--channel-list&knob-unread=true&viewMode=story'
     );
 
     const image = await page.screenshot();